Enable index files to be re-indexed i.e. updated
This was tricky at first because clucy’s delete func isn’t as straightforward
as one might think. For some reason — still not entirely clear to me — it
lower-cases all keys and values in the query it creates based on the input
map. So I realized I had to lower-case the specific fields I’d be using
to effect the delete (:omnia-source and :omnia-source-id). Then by using
Cursive’s interactive debugger I was able to discern that when Clucy transforms
a map entry into a Lucene field it attaches certain indexing instructions to
the field, and one of those instructions indicated whether the field value
should be analyzed/tokenized. It seemed to me that since I wanted to only
and always match on the entire scalar value, I should disable analysis/tokenization
for these two fields. So I did, and now it’s working — I can delete a specific
document.