Fix an indexing bug that lead to duplicate Dropbox docs
When I renamed `omnia-source-id` to `omnia-file-id` I accidentally
renamed it here to `omnia-account-id` because I forgot that
`omnia-source-id` had meant “the ID of this file in the source” rather
than “the ID of the source of this file”. So then Dropbox documents
weren’t being deleted from the index before they were updated, because
the `omnia-file-id` values for Dropbox docs are the paths, and they include
chars like `/`, and since they were being analyzed, they weren’t being found
by the delete operations.

This one was a real bugger to figure out! I would have saved a lot of time had I
had tests...