Change default search field
So that only what I *want* to be searched, will be
searched. Right now that’s the text and the name.

Clucy’s default behavior was to create and index a
field called `_content` that contained a stringified
version of the entire map — keys and values. So
a search for “omnia” would always match because
“omnia” was in paths and names and stuff.