Upgrade Dropbox Client SDK
All I wanted to do was to use the new getPathDisplay
method so I could get the mixed-case path for each
file when indexing it. Well — also I wanted to move
the path prep code (e.g. removing the file name from
the path) from the Web ns to this ns. Anyway so I had
to upgrade the SDK from a beta to a release version
to do so, and apparently they tightened up all sorts of
things so I had to make all sorts of minor tweaks to
update to this new version.