Update dependencies
Including a pretty major change: upgrading Elastisch from 2.x to
3.0.0-beta1 — this is important for compatibility with ElasticSearch 2+,
and it’s been ages since ElasticSearch 2.0 was released; the current
version as of this writing is 2.4.1. Unfortunately Elastisch as a project
seems nearly defunct; this beta release dates back to June 12th, and only
2 commits have been merged in the past month.

So I might need to either start contributing to Elastisch, or drop it.